app/requirements.txt
requests
faiss-cpu
google-genai
google-adk
streamlit
python-dotenv